create base submission job parent class
